virt-operator: move node placement code to its own package
