+++ pkg/virt-operator/resource/placement/placement_suite_test.go
@@ -1,0 +1,11 @@
+package placement_test
+
+import (
+	"testing"
+
+	"kubevirt.io/client-go/testutils"
+)
+
+func TestComponents(t *testing.T) {
+	testutils.KubeVirtTestSuiteSetup(t)
+}
